io_space: Provide dummy implementation on all archs
